requirements.txt
tensorflow==2.13.0
numpy==1.24.3
scipy==1.12.0
pandas==2.2.1
polars==1.12.0
statsmodels==0.14.4
sympy==1.13.1
xgboost==2.0.3
lightgbm==4.3.0
tabpfn==2.0.8
cupy-cuda11x==13.3.0
numba==0.60.0
scikit-learn==1.4.2
imbalanced-learn==0.12.3
category-encoders==2.6.4
mrmr-selection==0.2.8
shap==0.46.0
lime==0.2.0.1
lazypredict==0.2.12
opencv-python==4.9.0.80
scikit-image==0.24.0
pillow==10.2.0
tifffile==2024.8.30
openslide-python==1.3.1
pyvips==2.2.3
wsi-handling==0.16.2
staintools==2.1.2
geopandas==1.0.1
rasterio==1.4.1
shapely==2.0.2
pyproj==3.6.1
rtree==1.3.0
pyogrio==0.10.0
affine==2.4.0
matplotlib==3.8.2
seaborn==0.13.2
plotly==5.23.0
tqdm==4.66.2
joblib==1.4.2
h5py==3.10.0
requests==2.31.0
rich==13.9.4
tabulate==0.9.0
pyyaml==6.0.1